travel_recommendation_app/src/requirements.txt
streamlit
openai
duckduckgo-search
agno
python-dotenv